+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 
 ### Added
+- Zooming (by using the mouse wheel) is now enabled in Design View. You can also select multiple project items by pressing 
+  the Ctrl-key down and dragging the mouse.
 
 ### Changed